pyproject.toml
[project]
name = "wjecexampaperprocessor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "json-repair>=0.41.0",
    "mistralai>=1.6.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]

